requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
garminconnect==0.2.36
pydantic==2.7.4
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
